+++ continew-auto-sky/src/main/java/top/continew/admin/auto/sky/model/resp/TaskCamiDetailResp.java
@@ -32,7 +32,7 @@ import java.time.*;
  * task-cami详情信息
  *
  * @author wjh
- * @since 2025/03/29 20:49
+ * @since 2025/04/12 20:21
  */
 @Data
 @ExcelIgnoreUnannotated
@@ -54,14 +54,14 @@ public class TaskCamiDetailResp extends BaseDetailResp {
      */
     @Schema(description = "任务Id")
     @ExcelProperty(value = "任务Id")
-    private Integer taskId;
+    private Long taskId;
 
     /**
      * camiId
      */
     @Schema(description = "camiId")
     @ExcelProperty(value = "camiId")
-    private Integer camiId;
+    private Long camiId;
 
     /**
      * 卡密类型
